Include some window setup code
By default, the NW app opens a window at a size and position that is
probably some default, but not the default that we want.

This adds in a simple bit of code that sets the window position and
size, which will be handy during development by me on this machine but
will probably piss off others if they play with it.

We're also specifying that we need a specific version of node, which we
have been told should match the version that is bundled with NW; the
list of those can be found at: https://nwjs.io/versions.json