Include some sample tray icon support
This is a long fought battle, and in the end, no tray icon for us,
because reasons.

The tray code seen here comes from the official nw tray example, where
it works because it's using an older version of NW.js. Newer versions
fails to work the tray icon properly on Linux, and older versions seem
to have problems launching the express app.

To this I say, *sigh*.

So, the code is here but does nothing, so we can test it on other
platforms and maybe include it there. In theory this could maybe be
left in place on Linux, where it won't do anything for some/most people
but may work for others, assuming that doesn't cause problems.