Solve our problem with window flash and position
This is a small tweak that solves a few of our problems for us all in
one shot.

Our first issue solved is that we don't want to hard code the URL for
the main page, since we might need to change the port on the fly, and
in that case we need the URL that opens to be on a different port.

To solve this we set the manifest to not show the default window, and
then create our own after the fact that points at the correct URL.

Our second issue is related to window position being wrong, but it
transpires that you can set a window ID (even in the manifest) and
NW.js will automatically save and restore that window's position, so we
do that now too.

As a result of these changes, the client code is no longer present,
being done in the node context instead.