Remove the minor curse that is NW.js
After some testing, we can only get our desired tray icon if we use an
older version of NW because tray icons in Linux haven't worked for 2
years at this point (at least on Slack/XFCE).

In addition, older versions don't have the same niceties for launching
node scripts, which is a bit of a bummer.

The largest letdown is that NW appears flakey as hell on Linux (again,
at least on my box) and routinely fails to display a window even though
it's actually running, which is the sad maker.