Bring in some initial application code
After stalling for a couple of days playing with different technologies
for bundling applications and distributing them, this brings in some
actual core for doing anything application related at all.

For the time being this is heavily piggybacking on the code that has
been done in the ruinous addons project, just to get us going.