Include an initial configuration system
This uses convict as a configuration schema system for specifying the
configuration items that we're interested in, since there are going to
be a few such things that we need to configure.

For the moment this is just bare bones as we get rolling; more config
options will be added, and we also need to work out a configuration
file as well, and load that if it exists.