Include the concept of a config file and folder
This includes a couple of core concepts.

Firstly, we determine, in an OS specific way, what location we should
be putting our configuration and other files into, ensuring that it
exists if it is not already in place.

Secondly, we check that location to see if there is a configuration file
and, if there is, we load it as a part of the config startup.