Upgrade our config bootstrapping at startup
The configuration folder that we're using is going to have more content
in it than just a single configuration file (for example a location
to install bundles in by default).

This augments our configuration bootstrap to use jetpack to do a quick
recursive folder copy at startup if the config folder isn't present,
which allows us to quickly set up the entire structure easily.