Include winston application logging
Set up a logging system, sending data both to the console as well as to
a file, both transports being configurable via new configuration
settings.

We default the console logging to on and file logging to off, but the
sample configuration file turns both of them on by default.

A new log folder has been added to the configuration area in order to
store the logs that are generated, which is populated with a README
because, you know, git. Also, it reminds you how to make logs show up
there.