Include a bundle schema validator
In preparation for actually scanning for and loading bundles at startup
this sets up a validation routine that will allow us to verify that if
a package contains an appropriate manifest key that indicates that it is
a bundle of ours, that the structure is as expected.

This uses the awesome Joker library by Axel669.