Give the config schema an entry for bundle dir
This is a value that is used in a variety of places in the code, and so
it makes sense to have it as a well known configuration value rather
than having a bunch of redundant versions of things that stick the rest
of the path onto the config folder.

This also opens the door to having the bundle location be distinct from
the rest of the config location if desired.