Add in bundle manifest dependency validation
This extends the code that was previously written that locates all of
the possible bundles and loads/validates them so that it also validates
dependencies.

The validation ensures that all dependent bundle for every package
exists, and that it also has a version that satisfies the requirement.

Any bundle that has a missing/invalid dependency will be removed from
the list of known bundles; this action is also taken recursively so that
anything that depends on something that was removed is also itself
removed.