Validate that duplicate bundles are not possible
This adds in an extra bundle discovery validation that ensures that no
bundle appears more than once in the distinct locations in which we load
bundles from by removing any that are found from the list.

This needs to be a bit more robust since it assumes that a duplicate
will only happen twice.