Enhance our duplicate bundle detection code
This enhances the previous code by ensuring that no matter how many
times a bundle appears, it will be flagged as a duplicate, since
previously the first duplicate would remove the item from the list,
allowing a second duplicate to take it's place undetected.

We now leave items in the list, mark them as duplicated, and trim them
away at the end so that we can be more robust.