Implement a topographical sort to get load order
This uses the DAG (well really D?G, may be cyclic we're currently
cheating a little bit) to recursively determine the appropriate
dependent load order for all found bundles.

The result is the order in which bundles need to be loaded in order to
ensure that all dependencies are satisfied at load time.