Display warnings for missing additional bundles
The bundle discovery code finds all bundles in the bundle directory and
loads only those which are bundles, which is what we want.

On the other hand, we also have a configuration element for a list of
extra folders that are expected to contain bundles. Previously if they
did not, it was silently dropped, which is bad because that seems
intentional.

This small change causes a warning to be logged in that case as a hint
to why things might be missing unexpectedly.