Update to the latest version of Joker
During our testing of bundle loading code we uncovered a bug in Joker
that caused optional double nested objects to incorrectly be flagged as
not having a value despite being optional.

Since this happened during the stream and Axel is awesome, the problem
was fixed a short 5 minutes after it was disclosed, hooray!