Include a custom event for bundle load errors
We plan on having a bundle loader which throws errors on load so that we
can detect when we might not be able to load packages because one of
their dependencies did not load.

To that end this is a custom exception that we can throw that indicates
that we just need to display an error message.

In use we will use this so that we can distinguish between the
exceptions that we throw (like files being missing) from exceptions
that user code throws so that we can display stack traces only in
useful situations.