Flesh out the bones of our bundle loader
This scaffolds out most of the remaining required code for our bundles
to load. We have some comments in place to remind us what steps need to
happen, but largely it's just a validation that static files exist,
dynamic route creation to serve those files from the server, and the
import and launch of some user defined code.