Check for a self referencing dependency
Although we are going to add a check for dependency loops which would
ultimately kick out this situation, add in an explicit test for
checking if a bundle depends on itself.

The main reason for this is so that the error message can be cleaner
(arguably) because this is an outlier situation that the general loop
detector should not have to special case, and just having a message
that says dependency loop, while valid, is less informative.