Include an initial version of dep loop detection
This implements a simple check for dependency loops between found
bundles which does a depth first search and a visited node stack to
detect if a loop has been found, which can be reported.

Currently this is a simple version that finds only a single loop at a
time; this should be modified to scan the entire graph and return all
loops all at once, but that is a task for a later day. Instead, we just
keep checking for loops until there are none.

If a circular dependency loop is found, we remove the offending modules
from the graph. No extra checks are needed since the cyclic loop
includes not only the modules that are forming the loop, but the module
that leads into the loop.