Refactor and reorganize the bundle loader code
This takes the file that used to be the bundle load and splits it into a
resolver and a loader so that there is a better separation of concerns
in the code.

As a part of this, some functions changed names to better represent what
they do, now that they are technically a public interface of sorts.