Fix a rather silly design error in the schema
In actually implementing the ability to serve panel and graphics to the
front end code, it immediately becomes obvious that perhaps we we want
to serve assets too and not just a single page.

Since the original design forgot that because I am a dingus, that is now
fixed. The overall schema now includes optional keys for specifying
where in the bundle panels and graphics should be served from, with
sensible defaults if they are not present.

This allows us to still specify the location that is served, while still
being physically able to do it.