Start with the code that serves panel files
This fleshes out the first parts of the bundle loader, and in particular
allows us to serve out any panel files for those bundles which provide
them.

This is still a heavy work in progress, and observably it might possibly
be going wrong because it does not seem to be logging what we expect
it to.

This requires more finesse with the paths in use and how it generates
the URL's, (such as including `/bundle/` in the path so that it doesn't
accidentally conflict with built in content, but we're getting there.