Don't use jetpack.inspect to check existence
There was a call in the resolver that uses jetpack.inspect to determine
if paths are a specific type; however this function returns undefined,
which means that we need to run an extra test; did it return a value
AND did the value it returned indicate that the item was of the right
type?

For our purposes we only need to see if things exist or not, and
jetpack.exists() tells us in one shot existence and type, which is
easier to grok.