Finalize the first version of panel/graphic loads
This tweaks and cleans up the code from yesterday that sets up the
routes that we need in order to serve out our panels and graphics from
bundles that have them.

This is fundamentally similar to how it used to work but has a lot more
error checking than it did previously (yesterday there were just a lot
of comments on what we SHOULD do, now we ACTUALY do them).

The code is also parametrized so that we can use it for both panels and
graphics without duplication.