Implement the initial extension loader module
This extends the bundle loader so that it will dynamically import the
extension endpoint file if one is specified and, if successful, will
invoke the exported endpoint lifecycle function from inside of the
extension file to allow it to fully initialize.

Currently the initialization function does not take any arguments, but
it should.