Include a simple starter version API
As a final task for the day, this extends the module loader with the
actual API that we should be passing to bundle extension points when
they load.

Currently the API contains only a logger that is specific to the bundle
and its package manifest (which could probably be somewhat cleaned up,
all things being equal), but this can be augmented later.