Alter our icons and footer image
One of the meanings of Omphalos is "a rounded stone", this alters the
placeholder images we've been using to match that more.

In particular the footer image is now a round-ish rock instead of the
poop that is used in the ruinous addons, and the favicon changes from
my face (which, admittedly, is somewhat rock-like).

Ultimately this should be redesigned, perhaps with some sort of control
in there (since another definition describes it as a center, hub or
control center), say a button or similar.

That hasn't been done yet since trying to do that in a way that doesn't
look dumb and which reads well as a smaller scale seems problematic
for someone as art useless as me.