Include a sample bundle in the repository
As we're working on the implementation for the API, the dashboard and
tying everything together, we are going to need a bundle that exercises
those features.

It probably behooves us to have an example of one in the repo and use
that for testing, so that when other people look at the code they have
a sample to refer to.