Add an initial doc website with docusaurus
Here we're playing with using docusaurus for the documentation/website
for the project. This initial commit is a modified version of the
sample that you get if you do something like:

    npx create-docusaurus@latest website classic

In the root of the project, and then make a bunch of changes to it.

In particular, we updated the package.json to have arguments on the
commands that we use so that they work in our context, and we added a
missing postcss config that doesn't come with the template because the
devs of docusaurus want to make sure that the defaults works for people
in a monorepo that want a shared CSS config but not for people that
don't. Seems weird, but *shrug*.

We also deleted a bunch of files and renamed some content and such forth
to make it look like something we care about to start with.