Do some more initial documentation work
This adds some tweaks to the docusaurus portion of the project to
enhance layouts and put in a little bit more text, though there is
still a lot to do in this regard.

This was done off stream on the theory that most people are not super
interested in watching smoke come out of someone's ears while trying to
to figure out what to put into the documentation.