Include gridstack in the codebase
I honestly can't remember where I pulled these files from while I was
doing pre-devember testing, but this includes what is presumably the
latest versions of gridstack's JavaScript and CSS files.

As I recall from my pre-devember playtime, the way that this is packaged
in node make it nigh impossible to use as a package because it uses
some extremely wacky package format that none of the tooling I'm using
is happy with.