Include some initial gridstack code
This sets up the basics for doing things with gridstack, though some
more tweaking is required here.

In particular the grid holding area is letting us move things out of
its bounds when it should be constraining us (or we need to turn the
scrollbar back on) and some more styling is needed.