Flesh out the status bar icons we will need
With the removal of the footer, this brings the help link into the
navigation bar and also includes the other links that we will
eventually need for client side content, although those pages are not
currently ready yet and so they are just placeholders.

This is the change that made us realize that daisyui has weird issues
with tooltips on the edge of the screen.