Include a route for getting panel data
This sets up the route that we want to use to be able to fetch the list
of panels from the UI so that they can be displayed.

This should be grabbing the information from the loaded bundles, but as
a first crack this is just hard coding the data for test purposes.