Populate the dashboard using our panel API
Although the API is still using canned data, this fleshes out the code
path for populating the panels dynamically into the dashboard at load
time.

More work needs to be done here, such as a better loading text, and
the fact that we probably need a tab control of some sort and several
grids, since workspaces.