Fix an issue with loading bundles w/no extensions
Due to a major brain fart the extension loader was first trying to use
the extension file key to get at the full path of the file, and THEN
checking to see if it was valid.

This caused bundles that have no extensions to error out the loader.

We now properly check existence; the status text now also matches the
same pattern as used in the asset loader.