Allow the bundle loader to clean visited nodes
The depth first search that we use to determine the load order was
leaving a "visited" key in the module objects.

Since we are passing these to modules at their load time so that they
know their own manifest, we don't want those keys present.

This alters the load order recursive function to be able to know when
the top level call is coming off the stack so it can do the cleanup
prior to a final return.