Give loaded bundles a clone of their manifest
Previously when a bundle was loaded it got a reference to the loaded
manifest file; this is an issue for bundles with dependencies since they
will get references to the manifest objects of their dependencies, which
would allow either the parent or dependent to step on each other's toes
with manifest changes.

We now clone the manifest that we pass in to ensure that this will not
be an issue.