Modify the bundle loader processing
Previously the bundle loader would return a list of routers that contain
asset routes for loaded items.

We need the list of loaded bundles in our API calls, so this has been
changed such that the loader returns the bundle graph instead, and it
is now injected into all request objects so that API methods have direct
access to it.

The bundle loader will return the bundle DAG for loaded bundles only.

As a part of this change, it is now a fatal load error to have a missing
asset path if you declare that you have assets of that type (e.g. saying
you have panels but not having the path preset).

It is currently not an error to declare assets that don't exist (panels
or graphics) since this may be by design during initial development.