Include bundle name in panel API information
Each panel that we generate should have a distinct ID value that we can
reference it with. This change applies that by coming up with an ID that
is a combination of the name of the panel from the manifest and the name
of the bundle that it is contained within, which taken together should
be unique.