Add a placeholder for the workspace display
Our dashboard will be able to have multiple workspaces, which are
controlled by the bundles that we load.

This fleshes out what the UI will ultimately look like by adding in a
tab display of workspace names. This is currently just hard coded to a
single workspace name for simplicity, but will be expanded on later.