Alter the load contract on the bundle loader
Previously the bundle loader would insert the router for routing panel
and overlay requests into the returned manifests. However this causes
issues because we need to clone the manifest when giving it out to
loaded bundles, and you can't clone a router.

So, we now return a separate list of routers and bundles to keep them
separate, since the routers are really only needed at init time.