Add style/script content to served panel/graphics
In order to allow our panels and overlays to talk to each other and the
back end, we need to have some common JS in both of them to set up and
maintain the connections, as well as to allow API access.

This is accomplished by having the code that serves the files parse
them, modify their content accordingly to add in the required data, and
then send the file content away.

Doing this allows us to add not only a script tag for getting at the API
but also a static representation of our configuration, and some default
CSS used by both panels and overlays by default.

Documentation was also updated.