Add routes for panels/overlays even if missing
Originally the theory was that we don't want a specialized route for
serving up specific panels and overlays if they're missing. Turns out
that this is not a good idea overall.

The reason is that we always set up a catchall static route for assets,
and in doing so we're setting up being able to serve the missing
panel/overlay if it appears at runtime.

Now, we generate the route always; the handler is smart enough to either
serve the appropriate file OR generate a stub error page that lets
the user know that something is missing.

Once the page appears, it will be served on reload without having to
restart anything.