Finish up our initial iframe panels
With the grace of @axel669 and others in chat, we have constructed a
layout for panels which includes a grid so that the title bar remains
at the same height and the remainder of the content resizes, which
allows us to position iframes inside of our panels.

In addition this also solves a nascent problem that we knew we had which
is that iframes will grab mouse events while the mouse is over them and
give them to their own content.

We tried to solve this by reflecting interior events back out to the
parent page, but either we did it wrong or gridstack is doing something
that is not happy with that.

Our final solution for the problem is to have a blocking div pop into
view to block all events while panels are moving or resizing.