Include socket.io and associated boilerplate code
This commit includes the socket.io server and client packages, the
code needed to actually start the socket server, and enough config code
to allow for us to configure CORS, so that we can use online testing
tools and the like during development.

The documentation is also updated to know about the new config value.