Include in served assets their configuration info
This is a small change and probably needs to be revisited, but the idea
is that panels and overlays need to know about themselves, particularly
their bundle so that they can direct events and the like when we get to
that part.

For now we're glossing over the bundle aspect, but we still want the
config present.