Include a fleshed out version of our eventual api
After blood, sweat, tears and admonitions from Stoney_Eagle about
whether or not we are leaking sockets, we have a good idea of how we
want our event system to work now.

We will need to have the server direct traffic, so all events need to be
dispatched using a specific "protocol" that is an object that contains
both the resulting bundle (or everywhere) that the event is directed at
along with the actual payload.

It should not take a lot of glue code on the back end to get everything
up and running since we are most of the way there already.