Set up a rollup config for our client side API
Simplistically, this bundles together all of the required files and
spits out a single API file that is meant to be loaded from panels and
graphics.

For the time being this is just the file we were playing with last
night, but we've done some testing to verify that once we start
exporting symbols, they will appear inside of the required object in
the browser for code there to access.