Update the graphic schema to include a name field
As we're working on the API, we want to try and treat everything as
asset agnostic as possible. This requires that a minimum number of
fields be the same across all kinds of assets.

One of those fields is name, which was previously only in panels but it
needs to be in graphics too. We allow it to be there and backfill it
with the name of the file if missing so that downstream code does not
need to care.