Flesh out our initial API for messages
This covers sending messages only, but this portion of the API allows
for easily sending a specific message either to all assets in your
bundle, all assets in some other bundle, or all assets everywhere.

The client side API is set up for this (though the server is currently
not), and there is server side code to support it.

Still missing for a complete implementation of this is the ability to
listen for a targeted event (currently the assets just dump every event
they get) and a server side version for extension code to use.