Tweak up the API documentation a little bit
Since we are starting to flesh out the API now, I have taken the time
off stream to tweak up the API portion of the documentation a little
bit to structure it more.

This is still not complete (no example code, the objects do not have
their members documented, etc) but it's a good starting point.