Add logging to the client side code
Bring the same logging interface to the front end that we have at the
back end. This uses the same configuration information as the back end
code does, so that both log in the same formats and at the same levels.

While working on this, it was noted that winston gets very angry if you
don't configure any transports. We should fix this so that it doesn't 
cause problems if the user doesn't want either log, but for now we will
leave that be.

A note has been made in the documentation.