Clean up how the API does exported symbols
This was a little convoluted previously, requiring an object to be
passed through various parts of the loader in order to capture the list
of symbols that are exported by extensions on load.

We now store that in the omphalos API object, which we are passing
through anyway so that it can be included in bundles when they load.

This cleans many call points up and reduces cognitive load.