Implement message passing in the server code
This makes the code in bundle extensions also have the ability to
receive messages where previously it could only send them, since the
messages that arrive from the client are on sockets and get reflected
out.

There is currently an issue that we need to look into related to the
broadcast messages. I have a theory that in order to broadcast (if we
want to) we need to listen specifically on a broadcast channel to do
so.