Include a working sample of messaging working
In the most amazing rube goldberg mechanism to date in this project,
this demos that our message passing is working by having a button in
one panel that dispatches a message that gets caught by the server,
which responds by sending a message of a different type up, allowing
an end to end to occur.

In addition, one of the panels also listens to the message that the
server listens for and responds as well.