Clean up client side API and remove broadcasting
This started out a simple change to remove the broadcast functionality
since we don't actually want it (it is possible to do so already via
existing functionality).

In an act of both making sure things are nice for the user as well as
guarding against my own dumbness, all of the client side API's now use
assertion checking to verify that they were called correctly.

So, this both cleans up everything, adds assertions, and also removes
the broadcast API endpoint as well.