Include some documentation updates
This fleshes out a bit more the quick start portion of the documentation
site, particularly the items in the quick start section.

This is much more meaty than it was before but there is still much to
do. The Creating Bundles page has the beginnings of a tutorial on
making a bundle, but this is put on hold until we have a way to display
the graphics so that we can screenshot things.