Fix client side logging time stamp boo boo
The refactored time stamp code in the client side logging is very clean
and concise. It is also a function you should call and not just log the
value of, because one of those gives you a timestamp, and the other one
makes you look like an idiot while you're live on Twitch.