Add in a toast system to the dashboard
This implementation is taken from the ruinous addons, and will
ultimately be used both by the dashboard code as needed as well as in
the API to allow code to display toasts in the dashboard.

Currently the navbar is using it to point out that buttons don't do
anything, but that they eventually will.