Create an API for displaying toasts
This uses the previously added toast system and allows for any panel,
graphic or extension code to display a toast.

To do this, the dashboard connects to the back end directly with a
special distinctly hard coded bundle and listens for a toast message
arrived for that bundle.

The API calls direct the appropriately shaped event to the client side
server bundle.

This all works, but the way that the API is added to the dashboard is
a little hacky; we should be serving it the same way as we do the other
assets.