Set up some initial client side routing
This sets up our pages similar to the ruinous addons project to allow us
to have various pages and swap between them without causing back end
reloads for everything.

This is currently using a ginned up version of the SPA request code from
ruinous, which needs to be made smarter since it needs to insert code
into the HTML file.