Adjust the back end server for SPA requests
For the client side code changes we're about to put forth, the main
landing part of the client code is the /dashboard/ url; the server will
now redirect there if you go to the root.

Additionally we server only the top level client application routes as a
SPA and allow other requests through, so that errors for missing assets
will propagate properly.