Finalize some initial thoughts on client nav
This theorizes that there can be any number of workspaces, and so
the main app component will gather all of the workspaces that are known
and pass them to the nav bar and the dashboard code.

The dashboard component can determine based on URL if the slug is valid
and send the browser to a valid dashboard in that case.