Allow for proper 404 error pages using client code
We want a unified page experience, which includes error pages.
Previously the 404 page for missing links was the generic one for
express, which is retina burning to say the least.

This is a small but effective change to the recent code for serving
static SPA files that allows for setting the status code. This allows
us to handle a wildcard route as the last route which will load the
main SPA page, but with a 404 status.

The client side router will display the error page that you see client
side, so that the experience is consistent.