Clean up the boilerplate page code for the stub UI
Since we have 404 errors working now, this cleans up the 404 page to
have more useful content in it.

As a part of that change we use the same scheme to flesh out the content
of the other UI pages, which will have content that needs to be
displayed in a similar way.