Give asset configs a type
When the omphalos API object is created, there is an asset configuration
object included. This tiny change stamps each object with the type of
asset that the configuration is for, so that things that use the API
will know.

This allows for things like sharing code between overlays and graphics,
for example.