Update the graphic page to be able to copy URL's
This is a small bit of scaffolding to enable the button for copying the
URL of a graphic to the clipboard so that it actually does that.

This also fixes a minor z order issue with tooltips going under the
header buttons if there is more than one graphic on display.