Give a complete "fake" bundle to the dashboard
When we serve the dashboard page, we want it to have access to the API,
and so we need to give it the same information that we would give
panels and graphics.

Up until now this was a stub object, but it's now a more fully formed
and valid item.

As a part of this, the dashboard bundle has a dependency on each of the
loaded modules, since what it does indeed does depend on those bundles
providing things.