Finalize the first MVP version :confetti:
This finishes up the first draft of the graphics page, which now will
dynamically show all overlays in all bundles, allow you to view their
contents and copy their URL's to the clipboard.

The functionality for reloading some or all of the graphics is still not
done, but that is a not a showstopper for the first MVP version.