Fix toast event issues
The original implementation of the toast system had an issue in that the
item that listens for the event and displays the toast was only
initialized in the dashboard component.

This means that unless the dashboard is visible, other code can't
trigger a toast.

In addition, every time you leave the dashboard and come back, a
duplicate handler would be registered, which causes the toasts to
multiply like tribbles.