Fix an issue with triggering toasts from the dash
We noticed an issue in which trying to manually trigger a toast fro the
console while in the dashboard would do nothing; this is because when
it sends off the message, the server will not direct it back to us
since we sent it.

This is resolved by having, for the specific case of the toast() API
call, just raising the local event directly to do the same thing.

This also fixes a latent error in the toast mechanism in that doing math
on `undefined` nets you a `NaN`, which the toast library does not like.

This is really only an issue for manually triggering, but it's fixed
anyway because wrong is wrong.