Fix a theming issue
The application base for this project was taken from ruinous (or at
least the stub HTML files were) which causes theme issues.

In particular, in that app there is a theme selector and local storage
keeps the theme selected by the user. Here in this new app we don't
have that button, but we still use the local storage.

As a result, for anyone other than me, the theme ends up being light,
which looks very wrong.

Fixed now by forcing the theme to just always be dark. Thanks for this
go to @codebrainz for noticing the issue by being the first person
other than me to run the thing.