Add a system wide event listener for reloading
For an upcoming change, we need to be able to direct a reload request to
any asset of any type in any bundle.

Messages are already per bundle, so this implements a well known event
name and a handler for doing a reload based on a criteria. Every panel
and graphic that loads applies the listener, and the handler will use
the contents of the incoming request to know if it should reload or
not.