Include a reload button in panels
As a first item in developer mode (which was sneak previewed by the
change in the documentation, I just realized), this uses the new flag
to determine if it should add a reload button to panels.

These are useful during development since it lets you target a reload on
a specific panel without upsetting everything else.