Clarify the helper name for getting graphic info
In the code that displays all of the graphics, we're calling an internal
helper function that gathers all of the graphic information from the
loaded bundles given to us when the API loads.

This is a very minor change, but alter the name to be more indicative of
what it's actually doing.