Add persistence to the dashboard panel layout
This change brings in the ability for the system to remember on a workspace
by workspace basis the panel position.

This is done by hooking into the events that tell us when a drag or
resize operation is finished and grabbing the layout out to store into
local storage, and loading that back on load.

When a workspace loads, its panel configuration is augmented by any
saved information, with position and size information overlaying on top
if their values have changed.

As a part of this change, we no longer need the back end API to give us
the panel list because that is actually packed in as a part of a
previous change related to the API.