Add additional panel controls
This change updates the panel controls to provide more control to the
user regarding their dashboard layout.

There is now a pin button which tracks the state of whether or not a
panel can be automatically bumped out of the way by other panels when a
move/resize happens.

When pinned, the user can still move the panel around, but it will remain
in its assigned location and can't be nudged automatically.

There is now also a lock button, which tracks the state of being able to
move/resize the panel at all.

When locked, the user cannot move or resize a panel, but the system can
still move it out of the way if layout changes require more space.

The two controls are independent, so that in order to make the panel
fully "stuck" in place, both would need to be turned on.