Abstract our socket communications a bit
This sets the stage for a communications protocol in which all incoming
connections need to announce who they are in order to be able to
communicate, which will allow us to track exactly what connection is
for what, something needed for upcoming changes.

As a part of this change there is a small tweak to the square panel so
that it conveys its information better, and there are also some logging
changes.