Include a first version of client tracking code
This is a small change but it allows us to track on a socket by socket
basis who is a valid client and who is not based on them providing us an
appropriate "hello" message on connection.

The current protocol silently drops any messages from a socket unless it
is a HELLO message, requiring all clients to say who they are before
they can say what they want.