Swap the graphics list to use a store
Previously we were using a static array that is pulled from the packed
in bundle information available to the dashboard to come up with the
list of graphics.

This has been swapped to be a readable store instead, which opens the
door for us to be able to update load counts easily once we give the
back end the ability to give us that kind of information.

More work would be needed since the store would need to change, but the
boilerplate is in place regardless.