Include back end socket update notifications
This includes the code required to have the back end send an update to
the client side code whenever a socket connection changes state in a
way that causes the internal client list to change.

This is sent in the form of a message to the system bundle, allowing
anyone to listen for it if desired.