Set up for post devember-22
The repository for this project is now at
https://github.com/OdatNurd/omphalos instead of here; development will
continue at that repository, which is sadly not a fork of this one
because you can't fork a repo from an account its in into the same
account (go figure).

This commit represents the last of the changes here, to provide a little
bit of context for anyone that arrives at the site from an older link
(or, you know, if we send them there) to explain what's happening.